requirements.txt
asgiref==3.3.1
cfgv==3.3.1
distlib==0.3.4
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.10.1
djangorestframework==3.12.2
djangorestframework-simplejwt==5.0.0
filelock==3.4.2
gunicorn==20.1.0
identify==2.4.1
mccabe==0.6.1
nodeenv==1.6.0
Pillow==9.0.0
platformdirs==2.4.1
pre-commit==2.16.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2020.4
PyYAML==6.0
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.13.0
whitenoise==5.3.0
